Factor out shortcodes into modules
